fix(profiling): avoid casting hidden states to float32 before lm_head
lm_head weights are often bfloat16, passing float32 caused:
RuntimeError: expected mat1 and mat2 to have the same dtype, but got: float != c10::BFloat16

+++ scripts/run_profiling.py
@@ -230,7 +230,7 @@ def run_logit_lens(model, tokenizer, data_path, n_samples, device, dtype) -> dic
             for l in range(L):
                 if l not in cache.cache:
                     continue
-                h = cache.cache[l].float()  # (B, D)
+                h = cache.cache[l].to(dtype)  # (B, D)
                 # Apply final layer norm
                 if final_ln is not None:
                     try:
